+++ lib/view/page/book_detail/book_detail_page/book_detail_page.dart
@@ -10,6 +10,8 @@ import 'package:readme_app/core/constants/yh_style_icons.dart';
 import 'package:readme_app/model/book_detail_mock_data.dart';
 import 'package:readme_app/model/review_mock_data.dart';
 import 'package:readme_app/util/star_score/star_score.dart';
+import 'package:readme_app/view/page/book_detail/book_detail_page/components/book_detail_body.dart';
+import 'package:readme_app/view/page/book_detail/book_detail_page/components/book_detail_cover.dart';
 import 'package:readme_app/view/page/book_detail/book_detail_page/components/membership_button.dart';
 import 'package:readme_app/view/page/book_detail/book_detail_page/components/purchase_button.dart';
 
@@ -48,40 +50,7 @@ class _BookDetailPageState extends State<BookDetailPage>
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
       extendBodyBehindAppBar: true,
       appBar: _buildAppBar(context),
-      body: SingleChildScrollView(
-        child: Column(
-          children: [
-            _buildbookcover(context),
-            Padding(
-              padding: const EdgeInsets.all(10.0),
-              child: Column(
-                children: [
-                  _buildCategory(),
-                  SizedBox(height: 20),
-                  _buildTitle(),
-                  SizedBox(height: 8),
-                  _buildAuthorStore(),
-                  SizedBox(height: 20),
-                  _buildPrice(),
-                  Divider(
-                    thickness: 2,
-                  ),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    children: _buildTabBar(),
-                  ),
-                  Divider(
-                    thickness: 2,
-                  ),
-                  SizedBox(height: 15),
-                  _buildListContents(),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
+      body: BookDetailBody(),
       bottomSheet: Container(
         child: Row(
           mainAxisAlignment: MainAxisAlignment.center,
@@ -290,390 +259,12 @@ class _BookDetailPageState extends State<BookDetailPage>
     );
   }
 
-  Widget _buildListContents() {
-    return Padding(
-      padding: const EdgeInsets.all(10.0),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            "< 책 소개 >",
-            style: TextStyle(
-                color: Colours.app_sub_black,
-                fontSize: Dimens.font_sp20,
-                fontWeight: FontWeight.bold),
-          ),
-          SizedBox(height: 15),
-          Text(
-            "${bookDetail.bookInfo}",
-            style: TextStyle(
-              color: Colours.app_sub_black,
-              fontSize: Dimens.font_sp18,
-            ),
-            overflow: TextOverflow.ellipsis,
-            maxLines: 4,
-          ),
-          SizedBox(height: 20),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                "< 목차 >",
-                style: TextStyle(
-                  color: Colours.app_sub_black,
-                  fontSize: Dimens.font_sp20,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              SizedBox(height: 15),
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: List.generate(
-                  bookContentList.length,
-                  (index) {
-                    return Text(
-                      '- ${bookContentList[index].toString()}',
-                      style: TextStyle(
-                        fontSize: Dimens.font_sp18,
-                        color: Colours.app_sub_black,
-                      ),
-                    );
-                  },
-                ),
-              ),
-            ],
-          ),
-          SizedBox(height: 20),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                "< 작가 정보 >",
-                style: TextStyle(
-                  color: Colours.app_sub_black,
-                  fontSize: Dimens.font_sp20,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              SizedBox(height: 15),
-              Text(
-                "${bookDetail.author}",
-                style: TextStyle(
-                  color: Colours.app_sub_black,
-                  fontSize: Dimens.font_sp18,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              SizedBox(height: 10),
-              Text(
-                "${bookDetail.authorInfo}",
-                style: TextStyle(
-                  color: Colours.app_sub_black,
-                  fontSize: Dimens.font_sp18,
-                ),
-                overflow: TextOverflow.ellipsis,
-                maxLines: 4,
-              ),
-              SizedBox(height: 15),
-              Divider(
-                thickness: 2,
-              ),
-              SizedBox(height: 15),
-              Text(
-                "< 리뷰 관리 >",
-                style: TextStyle(
-                  color: Colours.app_sub_black,
-                  fontSize: Dimens.font_sp20,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              SizedBox(height: 15),
-              Column(
-                children: List.generate(
-                  riewList.length,
-                  (index) {
-                    return Container(
-                      margin: EdgeInsets.symmetric(vertical: 5),
-                      width: MediaQuery.of(context).size.width,
-                      decoration: BoxDecoration(
-                        border: Border.all(
-                          style: BorderStyle.solid,
-                          color: Colours.app_sub_grey,
-                        ),
-                        borderRadius: BorderRadius.all(
-                          Radius.circular(10),
-                        ),
-                      ),
-                      child: Padding(
-                        padding: const EdgeInsets.all(10.0),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Row(
-                              children: [
-                                StarScore(score: riewList[index].stars),
-                                Spacer(),
-                                Text("${riewList[index].userId} "),
-                                Text("|  ${riewList[index].writeTime}")
-                              ],
-                            ),
-                            SizedBox(height: 10),
-                            Text(riewList[index].content),
-                          ],
-                        ),
-                      ),
-                    );
-                  },
-                ),
-              ),
-              SizedBox(height: 15),
-              Divider(
-                thickness: 2,
-              ),
-              _buildRivewWrite(context),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
 
-  Container _buildRivewWrite(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.symmetric(vertical: 15),
-      width: MediaQuery.of(context).size.width,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Container(
-            margin: EdgeInsets.only(bottom: 15),
-            child: Text(
-              "< 리뷰 작성 >",
-              style: TextStyle(
-                color: Colours.app_sub_black,
-                fontSize: Dimens.font_sp20,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ),
-          Container(
-            decoration: BoxDecoration(
-              border: Border.all(
-                style: BorderStyle.solid,
-                color: Colours.app_sub_darkgrey,
-              ),
-              borderRadius: BorderRadius.all(
-                Radius.circular(10),
-              ),
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: RatingBar.builder(
-                    initialRating: _rating,
-                    minRating: 0,
-                    direction: Axis.horizontal,
-                    allowHalfRating: true,
-                    itemCount: 5,
-                    itemSize: 18,
-                    itemBuilder: (context, _) => Icon(
-                      Icons.star,
-                      color: Colors.amber,
-                    ),
-                    onRatingUpdate: (rating) {
-                      setState(() {
-                        _rating = rating;
-                      });
-                    },
-                  ),
-                ),
-                Padding(
-                  padding: const EdgeInsets.all(5.0),
-                  child: TextFormField(
-                    controller: _textController,
-                    decoration: InputDecoration(
-                      hintText: "내용을 입력하세요",
-                      border: OutlineInputBorder(),
-                    ),
-                    maxLines: 3,
-                  ),
-                ),
-                Padding(
-                  padding: const EdgeInsets.all(5.0),
-                  child: Align(
-                    alignment: Alignment.bottomRight,
-                    child: ElevatedButton(
-                      onPressed: () {
-                        // 등록 후 위치 이동
-                      },
-                      child: Text("등록"),
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colours.app_sub_black,
-                        foregroundColor: Colours.app_sub_white,
-                        textStyle: TextStyle(fontSize: Dimens.font_sp14),
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-          SizedBox(height: 60),
-        ],
-      ),
-    );
-  }
 
-  Row _buildPrice() {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.start,
-      children: [
-        Text(
-          "소장가 ${bookDetail.price}원",
-          style: TextStyle(
-            fontSize: Dimens.font_sp20,
-            color: Colours.app_sub_black,
-          ),
-        ),
-      ],
-    );
-  }
 
-  Row _buildAuthorStore() {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.start,
-      children: [
-        Text(
-          "${bookDetail.author}",
-          style: TextStyle(
-            fontSize: Dimens.font_sp24,
-            color: Colours.app_sub_black,
-          ),
-        ),
-        SizedBox(width: 10),
-        Text(
-          "/ ${bookDetail.store}",
-          style: TextStyle(
-            fontSize: Dimens.font_sp24,
-            color: Colours.app_sub_black,
-          ),
-        ),
-      ],
-    );
-  }
 
-  Row _buildTitle() {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.start,
-      children: [
-        Padding(
-          padding: const EdgeInsets.only(bottom: 5.0),
-          child: Text(
-            "${bookDetail.title}",
-            style: TextStyle(
-              fontSize: Dimens.font_sp30,
-              fontWeight: FontWeight.w700,
-              color: Colours.app_sub_black,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
 
-  Row _buildCategory() {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.start,
-      children: [
-        Container(
-          padding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 10.0),
-          decoration: BoxDecoration(
-            border: Border.all(
-              color: Colours.app_sub_darkgrey,
-              width: 3,
-            ),
-            borderRadius: BorderRadius.circular(20),
-          ),
-          child: Text(
-            "# ${bookDetail.bigCategory}",
-            style: TextStyle(
-              fontSize: Dimens.font_sp14,
-              color: Colours.app_sub_black,
-            ),
-          ),
-        ),
-        SizedBox(width: 10),
-        Container(
-          padding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 10.0),
-          decoration: BoxDecoration(
-            border: Border.all(
-              color: Colours.app_sub_darkgrey,
-              width: 3,
-            ),
-            borderRadius: BorderRadius.circular(20),
-          ),
-          child: Text(
-            "# ${bookDetail.smallCategory}",
-            style: TextStyle(
-              fontSize: Dimens.font_sp14,
-              color: Colours.app_sub_black,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
 
-  List<Widget> _buildTabBar() {
-    return [
-      TextButton(
-        style: TextButton.styleFrom(
-          textStyle: const TextStyle(
-            fontSize: Dimens.font_sp20,
-          ),
-        ),
-        onPressed: () {
-          setState(() {
-            _selectedButtonIndex = 0;
-          });
-        },
-        child: Text(
-          "도서 정보",
-          style: TextStyle(
-            color: _selectedButtonIndex == 0
-                ? Colours.app_sub_black
-                : Colours.app_sub_black,
-            fontSize: Dimens.font_sp20,
-            fontWeight:
-                _selectedButtonIndex == 0 ? FontWeight.bold : FontWeight.normal,
-          ),
-        ),
-      ),
-      TextButton(
-        style: TextButton.styleFrom(
-          textStyle: const TextStyle(fontSize: Dimens.font_sp20),
-        ),
-        onPressed: () {
-          setState(() {
-            _selectedButtonIndex = 1;
-          });
-        },
-        child: Text(
-          "리뷰 관리",
-          style: TextStyle(
-            color: _selectedButtonIndex == 1
-                ? Colours.app_sub_black
-                : Colours.app_sub_black,
-            fontSize: Dimens.font_sp20,
-            fontWeight:
-                _selectedButtonIndex == 1 ? FontWeight.bold : FontWeight.normal,
-          ),
-        ),
-      ),
-    ];
-  }
 
   AppBar _buildAppBar(BuildContext context) {
     return AppBar(
@@ -717,64 +308,6 @@ class _BookDetailPageState extends State<BookDetailPage>
     );
   }
 
-  Stack _buildbookcover(BuildContext context) {
-    return Stack(
-      children: [
-        ImageFiltered(
-          imageFilter: ImageFilter.blur(sigmaX: 10, sigmaY: 10),
-          child: CachedNetworkImage(
-            alignment: Alignment.center,
-            height: 470,
-            width: double.infinity,
-            placeholder: (context, url) => const CircularProgressIndicator(),
-            imageUrl:
-                'https://readmecorpbucket.s3.ap-northeast-2.amazonaws.com/bookcover/893f6a77-47be-452c-9b1a-148b79785e24-book1.png',
-            fit: BoxFit.fill,
-          ),
-        ),
-        Container(
-          height: 500,
-          width: 500,
-          child: Padding(
-            padding: EdgeInsets.all(75),
-            child: CachedNetworkImage(
-              alignment: Alignment.center,
-              placeholder: (context, url) => const CircularProgressIndicator(),
-              imageUrl:
-                  'https://readmecorpbucket.s3.ap-northeast-2.amazonaws.com/bookcover/893f6a77-47be-452c-9b1a-148b79785e24-book1.png',
-              fit: BoxFit.cover,
-            ),
-          ),
-        ),
-        Positioned(
-          left: 280,
-          top: 450,
-          child: InkWell(
-            onTap: () {
-              Navigator.pushNamed(context, "/viewer");
-            },
-            child: Container(
-              padding: EdgeInsets.symmetric(vertical: 8.0, horizontal: 15.0),
-              decoration: BoxDecoration(
-                border: Border.all(
-                  color: Colours.app_sub_white,
-                  width: 3,
-                ),
-                borderRadius: BorderRadius.circular(20),
-              ),
-              child: Text(
-                "미리보기",
-                style: TextStyle(
-                  fontSize: Dimens.font_sp16,
-                  color: Colours.app_sub_white,
-                ),
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
 }
